+++ infra/database.ts
@@ -1,0 +1,6 @@
+export const researchTable = new sst.aws.Dynamo("Research", {
+    fields: {
+        researchId: "string"
+    },
+    primaryIndex: {hashKey: "researchId"},
+})