+++ infra/frontend.ts
@@ -1,6 +1,6 @@
-import { bucket } from "./storage";
+import { api } from "./api";
 
 export const frontend = new sst.aws.Nextjs("MyWeb", {
+  link: [api],
   path: "packages/frontend",
-  link: [bucket],
 });